Hotfix/article crud

+++ api/articles/views/article_crud_views.py
@@ -3,6 +3,7 @@ from rest_framework.exceptions import PermissionDenied
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+from rest_framework.views import APIView
 
 from ..models import Article, ArticleImage
 from ..s3instance import S3Instance
@@ -13,19 +14,17 @@ from ..serializers import ArticleImageSerializer, ArticleSerializer
 class ArticleCreateView(generics.CreateAPIView):
     queryset = Article.objects.all()
     serializer_class = ArticleSerializer
-    permission_classes = [IsAuthenticated]
-    parser_classes = (MultiPartParser, FormParser)
+    permission_classes = [permissions.IsAuthenticated]
 
     def perform_create(self, serializer):
-        # tag_ids가 이미 리스트로 전달되면 이를 처리
         tag_ids = self.request.data.get("tag_ids", "")
         if isinstance(tag_ids, str):
             tag_ids = [
                 int(tag_id.strip()) for tag_id in tag_ids.split(",") if tag_id.strip()
             ]
 
-        if len(tag_ids) > 3:
-            raise serializers.ValidationError("태그는 최대 3개까지만 가능합니다.")
+        if len(tag_ids) > 1:
+            raise serializers.ValidationError("태그는 최대 1개까지만 가능합니다.")
 
         article = serializer.save(user=self.request.user, tag_ids=tag_ids)
 
@@ -41,7 +40,6 @@ class ArticleUpdateView(generics.UpdateAPIView):
     queryset = Article.objects.all()
     serializer_class = ArticleSerializer
     permission_classes = [permissions.IsAuthenticated]
-    parser_classes = (MultiPartParser, FormParser)
     lookup_field = "id"
 
     def get_object(self):
@@ -55,23 +53,55 @@ class ArticleUpdateView(generics.UpdateAPIView):
     def perform_update(self, serializer):
         article = serializer.save()
 
-        images_data = self.request.FILES.getlist("images")
+        response_data = ArticleSerializer(
+            article, context={"request": self.request}
+        ).data
+        return Response(response_data, status=status.HTTP_200_OK)
+
+
+# 게시글 이미지 생성, 수정
+class ArticleImageUploadView(APIView):
+    permission_classes = [permissions.IsAuthenticated]
+    parser_classes = (MultiPartParser, FormParser)
+
+    def post(self, request, id):
+        article = Article.objects.get(id=id)
+
+        # 권한 확인
+        if article.user != request.user:
+            return Response(
+                {"message": "이미지 업로드 권한이 없습니다."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
+        images_data = request.FILES.getlist("images")
         if images_data:
             s3instance = S3Instance().get_s3_instance()
-            # 기존 이미지 삭제
+
+            # 기존 이미지 삭제 (필요한 경우)
             for image in article.images.all():
                 S3Instance.delete_file(s3instance, image.image)
                 image.delete()
 
             # 새로운 이미지 업로드
+            uploaded_images = []
             image_urls = S3Instance.upload_files(s3instance, images_data, article.id)
+
             for index, image_url in enumerate(image_urls):
                 is_thumbnail = index == 0  # 첫 번째 이미지를 썸네일로 설정
-                ArticleImage.objects.create(
+                article_image = ArticleImage.objects.create(
                     article=article, image=image_url, is_thumbnail=is_thumbnail
                 )
+                uploaded_images.append(article_image)
 
-        return article
+            # 업로드된 이미지 정보를 직렬화하여 반환
+            serialized_images = ArticleImageSerializer(uploaded_images, many=True)
+            return Response(serialized_images.data, status=status.HTTP_200_OK)
+
+        return Response(
+            {"message": "업로드된 이미지가 없습니다."},
+            status=status.HTTP_400_BAD_REQUEST,
+        )
 
 
 # 게시글 삭제
